feat: criando a camada de contexto das transações
