+++ ignite-reactjs-2023/04-dt-money/src/App.tsx
@@ -2,12 +2,15 @@ import { ThemeProvider } from "styled-components";
 import { defaultTheme } from "./styles/themes/default";
 import { GlobalStyle } from "./styles/global";
 import { Transactions } from "./pages/Transactions";
+import { TransactionsProvider } from "./contexts/TransactionsContext";
 
 export function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <GlobalStyle />
-      <Transactions/>
+      <TransactionsProvider>
+        <Transactions/>
+      </TransactionsProvider>
     </ThemeProvider>
   )
 }
