+++ ignite-reactjs-2023/04-dt-money/src/contexts/TransactionsContext.tsx
@@ -1,0 +1,41 @@
+import { createContext, ReactNode, useEffect, useState } from "react";
+
+export interface Transaction{
+    id: number,
+    description: string,
+    type: "income" | "outcome",
+    price: number,
+    category: string,
+    createdAt: string
+}
+
+
+interface TransactionContextType{
+    transactions: Transaction[];
+}
+
+interface TransactionsProviderProps{
+    children: ReactNode;
+}
+
+export const TransactionsContext = createContext({} as TransactionContextType);
+
+export function TransactionsProvider({children} : TransactionsProviderProps){
+    const [transactions, setTransactions] = useState<Transaction[]>([]);
+
+    async function loadTransaction(){
+        const response = await fetch('http://localhost:3000/transactions');
+        const data = await response.json();
+
+        setTransactions(data);
+    }
+
+    useEffect(() => {
+        loadTransaction();
+    }, []);
+    return (
+        <TransactionsContext.Provider value={{transactions}}>
+            {children}
+        </TransactionsContext.Provider>
+    );
+}
