+++ ignite-reactjs-2023/04-dt-money/src/pages/Transactions/index.tsx
@@ -1,31 +1,12 @@
-import { useEffect, useState } from "react";
+import { useContext } from "react";
 import { Header } from "../../components/Header";
 import { Summary } from "../../components/Summary";
 import { SeachForm } from "./components/SearchForm";
 import { PriceHighlight, TransactionsContainer, TransactionsTable } from "./styles";
-
-interface Transction{
-    id: number,
-    description: string,
-    type: "income" | "outcome",
-    price: number,
-    category: string,
-    createdAt: string
-}
+import { TransactionsContext } from "../../contexts/TransactionsContext";
 
 export function Transactions(){
-    const [transactions, setTransactions] = useState<Transction[]>([]);
-
-    async function loadTransaction(){
-        const response = await fetch('http://localhost:3000/transactions');
-        const data = await response.json();
-
-        setTransactions(data);
-    }
-
-    useEffect(() => {
-        loadTransaction();
-    }, []);
+    const {transactions} = useContext(TransactionsContext);
 
     return (
         <div>
